refactor: CategoryGridItem has its own file
